Compensation for Better Filament 5 Loading
Idler Tilt Compensation:
Right side of the Idler Body sits 0.5, 0.8 or 1,4mm lower compared to the left side, to mitigate issue with loading filament 5, as some users have been reporting.
The older version of the Idler Body (C6) utilized a 0.3mm compensation which apparently might not be sufficient for some UltiMulti units.

This modification should result in higher idler tension for the filaments on the right side.

Needs further testing.